backend/atlas_project/requirements.txt
-ip==22.3
Django==4.1.4
PyJWT==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7